Refactor WbSearchEntitiesResult into an interface
